refactor: OrderValidator인터페이스를 만들고 TableGroupService에서참조
